requirements.txt
Django==5.1.8
djangorestframework==3.16.0
django-cors-headers==4.3.1
django-filter==24.2
django-allauth==0.61.1
django-debug-toolbar==4.3.0
django-storages==1.15.0
celery==5.3.6
redis==5.0.2
Pillow==10.2.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
gunicorn==22.0.0
whitenoise==6.6.0
dj-database-url==2.1.0
markdown==3.5.2
pyyaml==6.0.2
requests==2.32.3
selenium==4.31.0
beautifulsoup4==4.13.3
webdriver-manager==4.0.2
pandas==2.2.1 